hookify some logic
