+++ components/ResidentPage/CustomGPDetailsEditor.tsx
@@ -1,11 +1,12 @@
 import useWarnUnsavedChanges from 'hooks/useWarnUnsavedChanges';
-import { KeyboardEventHandler, useEffect, useRef } from 'react';
+import { KeyboardEventHandler, useRef } from 'react';
 import { GPDetails, Resident } from 'types';
 import { useResident } from 'utils/api/residents';
 import { DataRow } from './DataBlock';
 import s from './CustomAddressEditor.module.scss';
 import { Field, Form, Formik, FormikProps } from 'formik';
 import { residentSchema } from 'lib/validators';
+import useClickOutside from 'hooks/useClickOutside';
 
 const Error = ({ error }: { error?: string }) =>
   error ? (
@@ -72,20 +73,12 @@ type InnerProps = Props & FormikProps<FormValues>;
 const InnerForm = ({ onClose, errors, touched, submitForm }: InnerProps) => {
   const ref = useRef<HTMLFormElement>(null);
 
+  useClickOutside(ref, onClose);
+
   const handleKeyup: KeyboardEventHandler = (e) => {
     if (e.key === 'Escape') onClose();
     if (e.key === 'Enter') submitForm();
   };
-
-  useEffect(() => {
-    const handleClickOutside = (e: MouseEvent) => {
-      if (ref.current && e.target && !ref.current.contains(e.target as Node))
-        onClose();
-    };
-
-    return () =>
-      document.removeEventListener('click', handleClickOutside, true);
-  }, [onClose]);
 
   return (
     <Form className={s.form} onKeyUp={handleKeyup} ref={ref}>
